chore: move the block commit test to epoch_205 test module, use config to set a lower epoch transition height
