+++ .github/workflows/bitcoin-tests.yml
@@ -51,7 +51,7 @@ jobs:
           - tests::neon_integrations::filter_long_runtime_tx_integration_test
           - tests::neon_integrations::mining_transactions_is_fair
           - tests::epoch_205::test_dynamic_db_method_costs
-          - tests::neon_integrations::epoch_2_05_transition_empty_blocks
+          - tests::epoch_205::transition_empty_blocks
     steps:
       - uses: actions/checkout@v2
       - name: Download docker image
